requirements.txt
numpy >= 1.24.1
psycopg2
Pillow

crcmod
torch==2.0.0